Navigation refactor
